Prepare home-grown persistent metadata for merge with RA one
